Title: 사각형 채우기 3, Time: 82ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240925/사각형 채우기 3/rectangle-fill-3.java
@@ -2,13 +2,11 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.*;
-/**
- *  시간 : 80ms, 메모리 : 11,928KB
- */
+
 public class Main {
     static int n;
     static final int MOD = 1_000_000_007;
-    static int[][] dp;
+    static long[][] dp;
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
@@ -18,7 +16,7 @@ public class Main {
             System.out.println(2);
             System.exit(0);
         }
-        dp = new int[n+1][2];
+        dp = new long[n+1][2];
         dp[1][1] = 2;
         dp[2][1] = 7;
         dp[2][0] = 1;
